refactor: move feature flag service
